+++ frontend/src/app/App.vue
@@ -3,7 +3,7 @@
 </script>
 
 <template>
-  <div class="app-container relative bg-[#020617] flex flex-col h-full">
+  <div class="app-container relative bg-[#020617] flex flex-col h-full overflow-hidden">
     <div class="absolute top-[-20%] left-[-10%] w-[70%] h-[70%] bg-indigo-900/50 blur-[120px] rounded-full pointer-events-none mix-blend-screen" />
     <div class="absolute top-[-10%] right-[-20%] w-[60%] h-[60%] bg-blue-500/50 blur-[100px] rounded-full pointer-events-none mix-blend-screen" />
     <div class="absolute bottom-[-20%] left-[10%] right-[10%] h-[60%] bg-blue-600/50 blur-[130px] rounded-full pointer-events-none mix-blend-screen" />